Update README.md/fixed errors
There were some formatting errors in the ReadMe file, as well as parts that were written for the SpeedTest plugin (not the VideoDownloader plugin). There may still be a few more things to fix, but I've cleaned up some of the errors and updated the file for the VideoDownloader plugin.